+++ packages/expo-analytics-amplitude/CHANGELOG.md
@@ -4,6 +4,19 @@
 
 ### 🛠 Breaking changes
 
+- Renamed all methods to include the 'Async' suffix:
+  - `initialize` to `initializeAsync`
+  - `setUserId` to `setUserIdAsync`
+  - `setUserProperties` to `setUserPropertiesAsync`
+  - `clearUserProperties` to `clearUserPropertiesAsync`
+  - `logEvent` to `logEventAsync`
+  - `logEventWithProperties` to `logEventWithPropertiesAsync`
+  - `setGroup` to `setGroupAsync`
+  - `setTrackingOptions` to `setTrackingOptionsAsync`
+    ([#9212](https://github.com/expo/expo/pull/9212/) by [@cruzach](https://github.com/cruzach))
+- All methods now return a Promise. ([#9212](https://github.com/expo/expo/pull/9212/) by [@cruzach](https://github.com/cruzach))
+
+
 ### 🎉 New features
 
 ### 🐛 Bug fixes
@@ -20,8 +33,8 @@ _This version does not introduce any user-facing changes._
 
 ## 8.2.1 — 2020-05-29
 
-*This version does not introduce any user-facing changes.*
+_This version does not introduce any user-facing changes._
 
 ## 8.2.0 — 2020-05-27
 
-*This version does not introduce any user-facing changes.*
+_This version does not introduce any user-facing changes._